+++ R/snippet_setup.R
@@ -5,17 +5,26 @@
 #' `r lifecycle::badge("experimental")`
 #'
 #' Intelligent code completion is an experimental way to type NMproject code.
-#' This function modifies/creates `r.snippets`.
+#' This function modifies/creates `r.snippets`.  Needs to be run interactively.
+#' Will ask for user confirmation since snippets are an RStudio config setting
 #'
-#' @param force Logical.  The default is `FALSE` which will require user confirmation before editting `r.snippets`
+#' @param force Logical.  The default is `FALSE` which will require user
+#'   confirmation before editting `r.snippets`.
+#'
+#' @param snippet_path Character path to the `r.snippets` file.
 #'
 #' @return No return value, called for side effects.
 #'
 #' @export
 
-setup_nmproject_code_completion <- function(force = FALSE) {
+setup_nmproject_code_completion <- function(force = FALSE,
+                                            snippet_path = find_snippet_path()) {
 
-  snippet_path <- snippet_path()
+  if (!force) {
+    if (!is_rstudio()) return(invisible())
+    if (!interactive()) return(invisible())
+  }
+
   template_path <- system.file("extdata", "r.snippets", package = "NMproject")
 
   snippet_exists <- file.exists(snippet_path)
@@ -40,7 +49,6 @@ setup_nmproject_code_completion <- function(force = FALSE) {
     })
 
     if (all(matching_snippets)) return(invisible())
-
   }
 
   ## Check the user is OK to make the modification
@@ -49,9 +57,7 @@ setup_nmproject_code_completion <- function(force = FALSE) {
 
     if (!interactive()) stop("Needs to be run interactively with force = FALSE")
 
-    msg <- "This will modify your user level RStudio snippets settings:
-    {usethis::ui_path(snippet_path)}
-    Are you OK with this?"
+    msg <- "Do you want to modify your user level RStudio settings to enable NMproject code completion?"
 
     ans <- usethis::ui_yeah(msg, yes = "Yes", no = "Abort", shuffle = FALSE)
     if (!ans) usethis::ui_stop("Aborting")
@@ -84,12 +90,56 @@ setup_nmproject_code_completion <- function(force = FALSE) {
     })
 
     write(snippet_contents, file = snippet_path)
-    usethis::ui_done("Code completion ready, hit TAB while coding to use")
+    usethis::ui_done("Snippet file updated: {usethis::ui_path(snippet_path)}")
 
+    if (!is.null(rstudioapi::getActiveProject())) {
+      ans <- usethis::ui_yeah("RStudio needs to restart for changes to come into effect",
+                              yes = "Restart now", no = "I'll do it later", shuffle = FALSE)
+      if (ans) rstudioapi::openProject()
+    } else {
+      usethis::ui_todo("Restart RStudio (Session -> Quit Session) for changes to take effect")
+    }
   }
 }
 
+snippets_startup_message <- function() {
 
+  if (!is_rstudio()) return(invisible())
+  if (!interactive()) return(invisible())
+
+  snippet_path <- find_snippet_path()
+  template_path <- system.file("extdata", "r.snippets", package = "NMproject")
+
+  snippet_exists <- file.exists(snippet_path)
+
+  msg <- "To set up NMproject code completion: 'setup_nmproject_code_completion()'"
+
+  if (!snippet_exists) {
+    packageStartupMessage(msg)
+    return(invisible())
+  }
+
+  ## check to see if modification is needed.
+  template_contents <- readLines(template_path)
+  snippet_contents <- readLines(snippet_path)
+
+  last_line_blank <- snippet_contents[length(snippet_contents)]
+  last_line_blank <- grepl("^\\s*$", last_line_blank)
+  if (!last_line_blank) snippet_contents <- c(snippet_contents, "")
+
+  last_line_blank <- template_contents[length(template_contents)]
+  last_line_blank <- grepl("^\\s*$", last_line_blank)
+  if (!last_line_blank) template_contents <- c(template_contents, "")
+
+  matching_snippets <- sapply(c("new_nm", "child"), function(snippet_name) {
+    to_index <- snippet_index(snippet_name, snippet_contents)
+    from_index <- snippet_index(snippet_name, template_contents)
+    identical(snippet_contents[to_index], template_contents[from_index])
+  })
+
+  if (!all(matching_snippets)) packageStartupMessage(msg)
+
+}
 
 snippet_index <- function(name, contents) {
 
@@ -115,7 +165,7 @@ snippet_index <- function(name, contents) {
 }
 
 
-snippet_path <- function() {
+find_snippet_path <- function() {
 
   rstudio_version <- rstudioapi::getVersion()
   if (rstudio_version < "1.3") {